+++ packages/core-js/full/math/index.js
@@ -1,5 +1,6 @@
 'use strict';
 var parent = require('../../actual/math');
+// TODO: Remove from `core-js@4`
 require('../../modules/esnext.math.clamp');
 require('../../modules/esnext.math.deg-per-rad');
 require('../../modules/esnext.math.degrees');