+++ packages/core-js/full/number/index.js
@@ -4,5 +4,6 @@ var parent = require('../../actual/number');
 module.exports = parent;
 
 require('../../modules/es.object.to-string');
+require('../../modules/esnext.number.clamp');
 require('../../modules/esnext.number.from-string');
 require('../../modules/esnext.number.range');